// src/app/features/administracion/enums/tipo-dato-campo.enum.ts
export enum TipoDatoCampo {
    NUMERO = 'NUMERO',
    TEXTO = 'TEXTO',
    BOOLEANO = 'BOOLEANO',
    FECHA = 'FECHA',
    HORA = 'HORA',
    CORREO_ELECTRONICO = 'CORREO_ELECTRONICO',
    LISTA_MULTIPLE_SELECCION = 'LISTA_MULTIPLE_SELECCION',
    LISTA_UNICA_SELECCION = 'LISTA_UNICA_SELECCION',
}